Fix build and test failures from packed array optimizations.
R=jkummerow@chromium.org

Review URL: https://chromiumcodereview.appspot.com/10332317

git-svn-id: http://v8.googlecode.com/svn/branches/bleeding_edge@11638 ce2b1a6d-e550-0410-aec6-3dcde31c8c00

+++ src/x64/macro-assembler-x64.cc
@@ -4059,7 +4059,7 @@ void MacroAssembler::LoadTransitionedArrayMapConditional(
   movq(scratch, Operand(scratch,
                         Context::SlotOffset(Context::JS_ARRAY_MAPS_INDEX)));
 
-  size_t offset = expected_kind * kPointerSize +
+  int offset = expected_kind * kPointerSize +
       FixedArrayBase::kHeaderSize;
   cmpq(map_in_out, FieldOperand(scratch, offset));
   j(not_equal, no_map_match);
